+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/custom/CustomCommentRepository.java
@@ -1,6 +1,6 @@
 package com.project.snsserver.domain.board.repository.jpa.custom;
 
-import com.project.snsserver.domain.board.model.dto.CommentResponse;
+import com.project.snsserver.domain.board.model.dto.response.CommentResponse;
 
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
