+++ src/main/java/com/project/snsserver/domain/board/repository/jpa/custom/CustomPostImageRepository.java
@@ -1,6 +1,6 @@
 package com.project.snsserver.domain.board.repository.jpa.custom;
 
-import com.project.snsserver.domain.board.model.dto.PostImageResponse;
+import com.project.snsserver.domain.board.model.dto.response.PostImageResponse;
 
 import java.util.List;
