+++ src/main/java/com/project/snsserver/domain/board/service/PostService.java
@@ -1,6 +1,11 @@
 package com.project.snsserver.domain.board.service;
 
-import com.project.snsserver.domain.board.model.dto.*;
+import com.project.snsserver.domain.board.model.dto.request.EditPostRequest;
+import com.project.snsserver.domain.board.model.dto.response.EditPostResponse;
+import com.project.snsserver.domain.board.model.dto.response.PostDetailResponse;
+import com.project.snsserver.domain.board.model.dto.response.PostHashtagResponse;
+import com.project.snsserver.domain.board.model.dto.response.PostImageResponse;
+import com.project.snsserver.domain.board.model.dto.response.PostResponse;
 import com.project.snsserver.domain.member.model.entity.Member;
 
 import org.springframework.data.domain.Pageable;
