+++ src/main/java/com/project/snsserver/domain/member/controller/MemberController.java
@@ -18,19 +18,19 @@ import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
-import com.project.snsserver.domain.member.model.dto.LoginRequest;
-import com.project.snsserver.domain.member.model.dto.LoginResponse;
-import com.project.snsserver.domain.member.model.dto.LogoutRequest;
-import com.project.snsserver.domain.member.model.dto.MemberDetailResponse;
-import com.project.snsserver.domain.member.model.dto.ProfileResponse;
-import com.project.snsserver.domain.member.model.dto.ReissueTokenResponse;
-import com.project.snsserver.domain.member.model.dto.SendAuthCodeRequest;
-import com.project.snsserver.domain.member.model.dto.SignUpRequest;
-import com.project.snsserver.domain.member.model.dto.SignUpResponse;
-import com.project.snsserver.domain.member.model.dto.UpdateNicknameRequest;
-import com.project.snsserver.domain.member.model.dto.UpdatePasswordRequest;
-import com.project.snsserver.domain.member.model.dto.VerifyAuthCodeRequest;
-import com.project.snsserver.domain.member.model.dto.WithdrawRequest;
+import com.project.snsserver.domain.member.model.dto.request.LoginRequest;
+import com.project.snsserver.domain.member.model.dto.response.LoginResponse;
+import com.project.snsserver.domain.member.model.dto.request.LogoutRequest;
+import com.project.snsserver.domain.member.model.dto.response.MemberDetailResponse;
+import com.project.snsserver.domain.member.model.dto.response.ProfileResponse;
+import com.project.snsserver.domain.member.model.dto.response.ReissueTokenResponse;
+import com.project.snsserver.domain.member.model.dto.request.SendAuthCodeRequest;
+import com.project.snsserver.domain.member.model.dto.request.SignUpRequest;
+import com.project.snsserver.domain.member.model.dto.response.SignUpResponse;
+import com.project.snsserver.domain.member.model.dto.request.UpdateNicknameRequest;
+import com.project.snsserver.domain.member.model.dto.request.UpdatePasswordRequest;
+import com.project.snsserver.domain.member.model.dto.request.VerifyAuthCodeRequest;
+import com.project.snsserver.domain.member.model.dto.request.WithdrawRequest;
 import com.project.snsserver.domain.member.model.entity.Member;
 import com.project.snsserver.domain.member.service.MemberService;
 import com.project.snsserver.global.util.AuthMember;
