--- src/main/java/com/project/snsserver/domain/member/model/dto/LoginRequest.java
@@ -1,21 +1,0 @@
-package com.project.snsserver.domain.member.model.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginRequest {
-
-	@NotBlank(message = "이메일을 입력해주세요.")
-	private String email;
-
-	@NotBlank(message = "비밀번호를 입력해주세요.")
-	private String password;
-}
