+++ src/main/java/com/project/snsserver/domain/member/model/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.member.model.dto;
+package com.project.snsserver.domain.member.model.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,7 +11,10 @@ import javax.validation.constraints.NotBlank;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class WithdrawRequest {
+public class LoginRequest {
+
+	@NotBlank(message = "이메일을 입력해주세요.")
+	private String email;
 
 	@NotBlank(message = "비밀번호를 입력해주세요.")
 	private String password;
