+++ src/main/java/com/project/snsserver/domain/member/model/dto/request/WithdrawRequest.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.board.model.dto;
+package com.project.snsserver.domain.member.model.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,8 +11,8 @@ import javax.validation.constraints.NotBlank;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class EditCommentRequest {
+public class WithdrawRequest {
 
-	@NotBlank(message = "내용을 입력해주세요.")
-	private String content;
+	@NotBlank(message = "비밀번호를 입력해주세요.")
+	private String password;
 }
