+++ src/main/java/com/project/snsserver/domain/member/model/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.member.model.dto;
+package com.project.snsserver.domain.member.model.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
