+++ src/main/java/com/project/snsserver/domain/member/model/dto/response/MemberDetailResponse.java
@@ -1,4 +1,4 @@
-package com.project.snsserver.domain.member.model.dto;
+package com.project.snsserver.domain.member.model.dto.response;
 
 import com.project.snsserver.domain.member.type.Gender;
 import com.project.snsserver.domain.member.type.MemberRole;
