+++ src/main/java/com/project/snsserver/domain/member/repository/jpa/custom/CustomMemberRepository.java
@@ -1,6 +1,6 @@
 package com.project.snsserver.domain.member.repository.jpa.custom;
 
-import com.project.snsserver.domain.member.model.dto.MemberDetailResponse;
+import com.project.snsserver.domain.member.model.dto.response.MemberDetailResponse;
 
 public interface CustomMemberRepository {
