+++ src/main/java/com/project/snsserver/domain/member/service/FollowService.java
@@ -5,7 +5,7 @@ import java.util.Map;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
-import com.project.snsserver.domain.member.model.dto.FollowResponse;
+import com.project.snsserver.domain.member.model.dto.response.FollowResponse;
 import com.project.snsserver.domain.member.model.entity.Member;
 
 public interface FollowService {
