+++ src/main/java/com/project/snsserver/domain/member/service/MemberServiceImpl.java
@@ -4,7 +4,18 @@ import com.project.snsserver.domain.awss3.service.AwsS3Service;
 import com.project.snsserver.domain.board.repository.jpa.*;
 import com.project.snsserver.domain.mail.model.MailMessage;
 import com.project.snsserver.domain.mail.service.MailService;
-import com.project.snsserver.domain.member.model.dto.*;
+import com.project.snsserver.domain.member.model.dto.request.LoginRequest;
+import com.project.snsserver.domain.member.model.dto.request.LogoutRequest;
+import com.project.snsserver.domain.member.model.dto.request.SignUpRequest;
+import com.project.snsserver.domain.member.model.dto.request.UpdateNicknameRequest;
+import com.project.snsserver.domain.member.model.dto.request.UpdatePasswordRequest;
+import com.project.snsserver.domain.member.model.dto.request.VerifyAuthCodeRequest;
+import com.project.snsserver.domain.member.model.dto.request.WithdrawRequest;
+import com.project.snsserver.domain.member.model.dto.response.LoginResponse;
+import com.project.snsserver.domain.member.model.dto.response.MemberDetailResponse;
+import com.project.snsserver.domain.member.model.dto.response.ProfileResponse;
+import com.project.snsserver.domain.member.model.dto.response.ReissueTokenResponse;
+import com.project.snsserver.domain.member.model.dto.response.SignUpResponse;
 import com.project.snsserver.domain.member.model.entity.LogoutAccessToken;
 import com.project.snsserver.domain.member.model.entity.Member;
 import com.project.snsserver.domain.member.model.entity.MemberAuthCode;
